Refactor utility methods into parent class for reuse
